refactor(branding): move BrandSettingsRepository interface to application/usecases
Applies the v0.141.0 DIP template to branding: relocate the single
BrandSettingsRepository interface from branding/domain/repositories/
to consumer-side branding/application/usecases/. Empty domain/repositories/
dir deleted per CLAUDE.md DDD-gate.

Mechanical scope (no behavior change, build/test green per package):
- file relocated (git rename detected)
- 5 consumers updated to drop the domain/repositories import:
  - application/usecases/get_branding.go (1 site)
  - application/usecases/update_branding.go (1 site)
  - infrastructure/persistence/brand_settings_pg.go — aliased import
    switched to application/usecases; compile-time assertion preserved
    (var _ usecases.BrandSettingsRepository = (*BrandSettingsRepositoryPG)(nil))
  - interfaces/http/handlers/admin_handler_test.go (fakeRepo)
  - interfaces/http/handlers/public_handler_test.go (stubRepo)
  - interfaces/http/routes/routes_test.go (fakeRepo)

Second module released under Phase 6 #196 multi-release sprint
(precedent: v0.143.0 announcements). Coverage backfill для usecases
follows in the next commit.
