+++ internal/modules/branding/application/usecases/get_branding.go
@@ -4,19 +4,18 @@ import (
 	"context"
 
 	"github.com/inf-sys-secretary-methodologist/inf-sys-secretary-methodist/internal/modules/branding/domain/entities"
-	"github.com/inf-sys-secretary-methodologist/inf-sys-secretary-methodist/internal/modules/branding/domain/repositories"
 )
 
 // GetBrandingUseCase reads the singleton brand settings.
 // Mounted under both the admin and the public HTTP groups —
 // no field is sensitive so the same projection serves both.
 type GetBrandingUseCase struct {
-	repo repositories.BrandSettingsRepository
+	repo BrandSettingsRepository
 }
 
 // NewGetBrandingUseCase constructs the use case. Panics on nil
 // repo so misconfigured DI fails at construction.
-func NewGetBrandingUseCase(repo repositories.BrandSettingsRepository) *GetBrandingUseCase {
+func NewGetBrandingUseCase(repo BrandSettingsRepository) *GetBrandingUseCase {
 	if repo == nil {
 		panic("branding: nil BrandSettingsRepository")
 	}
